+++ ttnn/cpp/ttnn/operations/conv/conv2d/device/kernels/conv_reader_common.hpp
@@ -8,7 +8,6 @@
 #include <cstdint>
 #include <tt-metalium/constants.hpp>
 #include "dataflow_api.h"
-#include "tt-metalium/math.hpp"
 
 // Zero out all tiles for a given circular buffer.
 template <uint32_t cb_id>
@@ -474,92 +473,4 @@ FORCE_INLINE void signal_write_done(volatile tt_l1_ptr uint32_t* semaphore_addr_
 FORCE_INLINE void wait_write_done(volatile tt_l1_ptr uint32_t* semaphore_addr_ptr) {
     noc_semaphore_wait(semaphore_addr_ptr, VALID);
     noc_semaphore_set(semaphore_addr_ptr, INVALID);
-}
-// Multicast activation data from the local circular buffer to multiple destinations.
-// This function sends a block of data (the activation block) using NOC multicast commands.
-//
-// Template parameters:
-//   act_mcast_num_cores: Number of core destinations in the multicast group.
-template <uint32_t act_mcast_num_cores>
-void multicast_data(
-    bool is_receiver_core,        // Whether this core is also a receiver of its own multicast (for loopback).
-    uint32_t src_l1_addr,         // SRC address in L1
-    uint32_t dst_l1_addr,         // DST address in L1
-    uint64_t multicast_noc_addr,  // Multicast NOC address
-    uint32_t total_bytes)         // Total bytes to send
-{
-    uint64_t multicast_write_addr = multicast_noc_addr | dst_l1_addr;
-    if (is_receiver_core) {
-        if constexpr (act_mcast_num_cores) {
-            // num_dests will source, since we are copying to a different local CB as well
-            noc_async_write_multicast_loopback_src(
-                src_l1_addr, multicast_write_addr, total_bytes, act_mcast_num_cores + 1, true);
-        } else {
-            // In this case sender core is the only reciever in the grid,
-            // we can't use the multicast_loopback_src (hang)
-            noc_async_write(get_noc_addr(src_l1_addr), get_noc_addr(dst_l1_addr), total_bytes);
-        }
-    } else {
-        // If sender core is not the reciever core as well we can't use the loopback mcast. (hang)
-        noc_async_write_multicast(src_l1_addr, multicast_write_addr, total_bytes, act_mcast_num_cores + 1, true);
-    }
-}
-
-template <
-    uint32_t act_mcast_num_dest_cores,
-    uint32_t mcast_noc_burst_size,
-    uint32_t block_tile_count,
-    uint32_t tile_size>
-void mcast_block(bool is_receiver_core, uint32_t src_cb, uint32_t dst_cb, uint64_t multicast_noc_addr) {
-    // number of full bursts
-    constexpr uint32_t mcast_full_burst_cnt = block_tile_count * tile_size / mcast_noc_burst_size;
-    // size of the leftover burst, if 0 means we have no leftover burst
-    constexpr uint32_t mcast_leftover_burst_size = block_tile_count * tile_size % mcast_noc_burst_size;
-    // number of tiles that we need to wait for to cover the full burst size
-    constexpr uint32_t wait_tile_full_cnt = (mcast_noc_burst_size + tile_size - 1) / tile_size;
-
-    // In full burst iterations we wait for a bit more than the full burst size in case where the
-    // tile size does not divide the burst size evenly.
-    // we need to insure that we don't wait for more tiles than we have in the block
-    constexpr uint32_t wait_tile_full_done = std::min(mcast_full_burst_cnt * wait_tile_full_cnt, block_tile_count);
-
-    // optimization to avoid unnecessary branching in the loop
-    constexpr bool no_need_partial_wait_tile = mcast_full_burst_cnt * wait_tile_full_cnt <= block_tile_count;
-
-    // number of times we need to increase the wait_tile_curr for the full burst iterations
-    constexpr uint32_t wait_tile_full_iter_cnt = (wait_tile_full_done / wait_tile_full_cnt) - 1;
-
-    uint32_t src_l1_addr = get_read_ptr(src_cb);
-    uint32_t dst_l1_addr = get_write_ptr(dst_cb);
-
-    constexpr uint32_t wait_tile_start_cnt = std::min(block_tile_count, wait_tile_full_cnt);
-    uint32_t wait_tile_curr = wait_tile_start_cnt;
-    for (uint32_t i = 0; i < mcast_full_burst_cnt; i++) {
-        cb_wait_front(src_cb, wait_tile_curr);
-        multicast_data<act_mcast_num_dest_cores>(
-            is_receiver_core, src_l1_addr, dst_l1_addr, multicast_noc_addr, mcast_noc_burst_size);
-        src_l1_addr += mcast_noc_burst_size;
-        dst_l1_addr += mcast_noc_burst_size;
-
-        if constexpr (no_need_partial_wait_tile) {
-            wait_tile_curr += wait_tile_full_cnt;
-        } else {
-            // we shouldn't wait for more than the number of tiles in the block
-            if (i < wait_tile_full_iter_cnt) {
-                wait_tile_curr += wait_tile_full_cnt;
-            } else {
-                wait_tile_curr = block_tile_count;
-            }
-        }
-    }
-    if constexpr (mcast_leftover_burst_size > 0) {
-        cb_wait_front(src_cb, block_tile_count);
-        multicast_data<act_mcast_num_dest_cores>(
-            is_receiver_core, src_l1_addr, dst_l1_addr, multicast_noc_addr, mcast_leftover_burst_size);
-    }
-    if constexpr (act_mcast_num_dest_cores == 0) {
-        if (is_receiver_core) {
-            noc_async_write_barrier();
-        }
-    }
 }
